Release: 0.15.2

+++ src/Telegram/Types/Sticker/MaskPosition.php
@@ -18,16 +18,16 @@ class MaskPosition
      * Shift by X-axis measured in widths of the mask scaled to the face size, from left to right.
      * For example, choosing -1.0 will place mask just to the left of the default mask position.
      */
-    public double $x_shift;
+    public float $x_shift;
 
     /**
      * Shift by Y-axis measured in heights of the mask scaled to the face size, from top to bottom.
      * For example, 1.0 will place the mask just below the default mask position.
      */
-    public double $y_shift;
+    public float $y_shift;
 
     /**
      * Mask scaling coefficient. For example, 2.0 means double size.
      */
-    public double $scale;
+    public float $scale;
 }
